Fix unittest and style improvement

+++ vantage6-server/tests_server/test_resources.py
@@ -5,6 +5,7 @@ import json
 import uuid
 import random
 import string
+from vantage6.server.default_roles import DefaultRole
 import yaml
 
 from http import HTTPStatus
@@ -386,6 +387,7 @@ class TestResources(unittest.TestCase):
         self.assertEqual(node.status_code, 404)
 
     def test_result_with_id(self):
+        # TODO improve this unit test to create its own task and run
         headers = self.login("root")
         run = self.app.get("/api/run/1", headers=headers)
         self.assertEqual(run.status_code, 200)
@@ -403,11 +405,6 @@ class TestResources(unittest.TestCase):
 
         result3 = self.app.get("/api/run?task_id=1", headers=headers)
         self.assertEqual(result3.status_code, 200)
-
-    def test_stats(self):
-        headers = self.login("root")
-        result = self.app.get("/api/run", headers=headers)
-        self.assertEqual(result.status_code, 200)
 
     def test_task_with_id(self):
         headers = self.login("root")
@@ -704,8 +701,9 @@ class TestResources(unittest.TestCase):
         rule_ids = [rule.get("id") for rule in rules]
 
         # assign first two rules to role
+        ROLE_TO_CREATE_NAME = "some-role-name"
         body = {
-            "name": "some-role-name",
+            "name": ROLE_TO_CREATE_NAME,
             "description": "Testing if we can create a role",
             "rules": rule_ids[:2],
         }
@@ -728,7 +726,8 @@ class TestResources(unittest.TestCase):
 
         # cleanup
         for resource in Role.get():
-            resource.delete()
+            if resource.name == ROLE_TO_CREATE_NAME:
+                resource.delete()
 
     def test_create_role_as_root_for_different_organization(self):
         headers = self.login("root")
@@ -742,8 +741,9 @@ class TestResources(unittest.TestCase):
         org = Organization(name="Some-random-organization")
         org.save()
 
+        ROLE_TO_CREATE_NAME = "some-role-name"
         body = {
-            "name": "some-role-name",
+            "name": ROLE_TO_CREATE_NAME,
             "description": "Testing if we can create a rol for another org",
             "rules": [rule.get("id") for rule in rules],
             "organization_id": org.id,
@@ -760,7 +760,8 @@ class TestResources(unittest.TestCase):
 
         # cleanup
         for resource in Role.get():
-            resource.delete()
+            if resource.name == ROLE_TO_CREATE_NAME:
+                resource.delete()
 
     def test_create_role_permissions(self):
         all_rules = Rule.get()
@@ -768,8 +769,9 @@ class TestResources(unittest.TestCase):
         # check user without any permissions
         headers = self.create_user_and_login()
 
+        ROLE_TO_CREATE_NAME = "some-role-name"
         body = {
-            "name": "some-role-name",
+            "name": ROLE_TO_CREATE_NAME,
             "description": "Testing if we can create a role for another org",
             "rules": [rule.id for rule in all_rules],
         }
@@ -837,7 +839,8 @@ class TestResources(unittest.TestCase):
 
         # cleanup
         for resource in Role.get():
-            resource.delete()
+            if resource.name == ROLE_TO_CREATE_NAME:
+                resource.delete()
 
     def test_edit_role(self):
         headers = self.login("root")
@@ -1820,17 +1823,19 @@ class TestResources(unittest.TestCase):
         self.assertEqual(result.status_code, HTTPStatus.UNAUTHORIZED)
 
         # create an organization
+        ORG_CREATED_NAME = "this-is-gonna-happen"
         rule = Rule.get_by_("organization", Scope.GLOBAL, Operation.CREATE)
         headers = self.create_user_and_login(rules=[rule])
         result = self.app.post(
-            "/api/organization", headers=headers, json={"name": "this-is-gonna-happen"}
+            "/api/organization", headers=headers, json={"name": ORG_CREATED_NAME}
         )
         self.assertEqual(result.status_code, HTTPStatus.CREATED)
-        self.assertIsNotNone(Organization.get_by_name("this-is-gonna-happen"))
+        self.assertIsNotNone(Organization.get_by_name(ORG_CREATED_NAME))
 
         # cleanup
         for resource in Organization.get():
-            resource.delete()
+            if resource.name == ORG_CREATED_NAME:
+                resource.delete()
 
     def test_patch_organization_permissions(self):
         # unknown organization
@@ -2802,7 +2807,8 @@ class TestResources(unittest.TestCase):
         org3.delete()
         org4.delete()
         col.delete()
-        for resource in Node.get():
+        # delete the last three nodes (i.e. those that were created in this test)
+        for resource in Node.get()[::-1][:3]:
             resource.delete()
 
     def test_delete_node_permissions(self):
@@ -3204,14 +3210,14 @@ class TestResources(unittest.TestCase):
         self.assertEqual(results.status_code, HTTPStatus.BAD_REQUEST)
 
         # cleanup
+        # delete the 1 task that was created in this unit test
+        Task.get()[::-1][0].delete()
         node.delete()
         org.delete()
         org2.delete()
         col.delete()
         col2.delete()
         study.delete()
-        for resource in Task.get():
-            resource.delete()
 
     def test_create_task_permissions_as_container(self):
         org = Organization()
@@ -3299,8 +3305,8 @@ class TestResources(unittest.TestCase):
         col.delete()
         node2.delete()
         col2.delete()
-        for resource in Task.get():
-            resource.delete()
+        # delete the 1 task that was created in this unit test
+        Task.get()[::-1][0].delete()
 
     def test_delete_task_permissions(self):
         # test non-existing task
@@ -3557,8 +3563,8 @@ class TestResources(unittest.TestCase):
         # cleanup
         org.delete()
         col.delete()
-        for store in AlgorithmStore.get():
-            store.delete()
+        for resource in AlgorithmStore.get():
+            resource.delete()
 
     def test_view_algorithm_store(self):
         """Test viewing algorithm store records"""
